Fix bugs

+++ 3lab/include/utils.h
@@ -1,6 +1,6 @@
 #include "../include/transport.h"
 #pragma once
 
-double validation();
+int validation();
 double validateWeight(double weight, double maxCapacity);
 int validatePassengers(int passengers, int maxPassengers);
